requirements.txt
dj-database-url==0.5.0
Django==2.0.7
django-cors-headers==2.4.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
pytz==2018.5
psycopg2==2.7.2